rename cli to _cli
