+++ src/stpipe/__main__.py
@@ -1,4 +1,4 @@
-from .cli.main import main
+from ._cli.main import main
 
 # main() always raises SystemExit, so we don't need
 # to perform any exception handling here.
