+++ src/stpipe/_cli/command.py
