+++ src/stpipe/_cli/list.py
