+++ src/stpipe/_cli/main.py
