+++ tests/cli/test_list.py
@@ -1,7 +1,7 @@
 import pytest
 
 from stpipe import entry_points
-from stpipe.cli import handle_args
+from stpipe._cli import handle_args
 from stpipe.entry_points import StepInfo
