+++ tests/cli/test_main.py
@@ -4,7 +4,7 @@ import pytest
 
 import stpipe
 from stpipe import entry_points
-from stpipe.cli import handle_args
+from stpipe._cli import handle_args
 from stpipe.entry_points import StepInfo
